+++ src/modules/utils/motordrivercontrol/drivers/TMC22X/TMC22X.cpp
@@ -1033,7 +1033,7 @@ void TMC22X::setStealthChopthreshold(uint32_t threshold)
     }
 }
 
-void TMC22X::set_current(unsigned int current)
+void TMC22X::set_current(uint16_t current)
 {
     uint8_t current_scaling = 0;
     //calculate the current scaling from the max current setting (in mA)
